// frontend/src/app/Modulos/ventas/pages/vendedores-rv/interface/VentasVendedores.interface.ts
export default interface VentasVendedores {
    Calculated_SellerName: string;
    Calculated_IDVendedor: string;
    Calculated_Venta:      number;
    Calculated_Costo:      number;
    Calculated_Peso:       number;
    Calculated_PPTO:       number;
    Calculated_Ejecutado:  number;
    Calculated_Pendiente:  number;
    Calculated_GAP:        number;
    Calculated_MG:         number;
    Calculated_MGp:        number;
    RowIdent:              string;
}